git status

+++ src/com/mariangolea/fintracker/banks/pdfparser/api/BankTransaction.java
@@ -10,95 +10,99 @@ import java.util.Objects;
  *
  * @author mariangolea@gmail.com
  */
-public final class BankTransaction implements Serializable{
+public final class BankTransaction implements Serializable {
 
-    public enum Type {
-        IN,
-        OUT
-    }
+	public enum Type {
+		IN, OUT
+	}
 
-    //Date when the transaction has been initiated by a client.
-    public final Date startDate;
-    //Date when the transaction has been settled by the bank. Same as startDate in most cases.
-    public final Date completedDate;
-    //amount of currency used in this transaction.
-    public final double amount;
-    //description of the transaction. useful for a client to recategorize a certain transaction.
-    public final String description;
-    //type of transaction.
-    public final Type type;
-    //title as it appears in pdf
-    public final String title;
+	// Date when the transaction has been initiated by a client.
+	public final Date startDate;
+	// Date when the transaction has been settled by the bank. Same as startDate in
+	// most cases.
+	public final Date completedDate;
+	// amount of currency used in this transaction.
+	public final float amount;
+	// description of the transaction. useful for a client to recategorize a certain
+	// transaction.
+	public final String description;
+	// type of transaction.
+	public final Type type;
+	// title as it appears in pdf
+	public final String title;
 
-    /**
-     * Constructs a transaction instance.
-     *
-     * @param title
-     * @param startDate date when transaction has been initiated by the client.
-     * @param completedDate date of transaction settlement.
-     * @param amount currency amount
-     * @param description transaction description (useful for a client to
-     * re categorize a certain transaction).
-     * @param type transaction type
-     */
-    public BankTransaction(final String title, final Date startDate, final Date completedDate, double amount, final String description, final Type type) {
-        super();
-        this.startDate = startDate;
-        this.completedDate = completedDate;
-        this.amount = amount;
-        this.description = description;
-        this.type = type;
-        this.title = title;
-    }
+	/**
+	 * Constructs a transaction instance.
+	 *
+	 * @param title
+	 * @param startDate
+	 *            date when transaction has been initiated by the client.
+	 * @param completedDate
+	 *            date of transaction settlement.
+	 * @param amount
+	 *            currency amount
+	 * @param description
+	 *            transaction description (useful for a client to re categorize a
+	 *            certain transaction).
+	 * @param type
+	 *            transaction type
+	 */
+	public BankTransaction(final String title, final Date startDate, final Date completedDate, float amount,
+			final String description, final Type type) {
+		super();
+		this.startDate = startDate;
+		this.completedDate = completedDate;
+		this.amount = amount;
+		this.description = description;
+		this.type = type;
+		this.title = title;
+	}
 
-    @Override
-    public String toString() {
-        String string = "";
+	@Override
+	public String toString() {
+		String string = "";
 
-        string = string.concat(title).concat(",");
-        string = string.concat(type.name()).concat(",");
-        string = string.concat(description).concat(",");
-        string = string.concat(DateFormat.getInstance().format(startDate)).concat(",");
-        string += amount;
+		string = string.concat(title).concat(",");
+		string = string.concat(type.name()).concat(",");
+		string = string.concat(description).concat(",");
+		string = string.concat(DateFormat.getInstance().format(startDate)).concat(",");
+		string += amount;
 
-        return string;
-    }
+		return string;
+	}
 
-    public String toString(final DateFormat customFormat) {
-        String string = "";
+	public String toString(final DateFormat customFormat) {
+		String string = "";
 
-        string = string.concat(title).concat(",");
-        string = string.concat(type.toString()).concat(",");
-        string = string.concat(description).concat(",");
-        string = string.concat(customFormat.format(startDate)).concat(",");
-        string += amount;
+		string = string.concat(title).concat(",");
+		string = string.concat(type.toString()).concat(",");
+		string = string.concat(description).concat(",");
+		string = string.concat(customFormat.format(startDate)).concat(",");
+		string += amount;
 
-        return string;
-    }
+		return string;
+	}
 
-    @Override
-    public boolean equals(Object obj) {
-        if (null == obj || !(obj instanceof BankTransaction)){
-            return false;
-        }
-        BankTransaction other = (BankTransaction) obj;
-        return title.equals(other.title)
-        		&& type == other.type 
-                && amount == other.amount
-                && startDate.equals(other.startDate) 
-                && completedDate.equals(other.completedDate) 
-                && description.equals(other.description);
-    }
+	@Override
+	public boolean equals(Object obj) {
+		if (null == obj || !(obj instanceof BankTransaction)) {
+			return false;
+		}
+		BankTransaction other = (BankTransaction) obj;
+		return title.equals(other.title) && type == other.type && amount == other.amount
+				&& startDate.equals(other.startDate) && completedDate.equals(other.completedDate)
+				&& description.equals(other.description);
+	}
 
-    @Override
-    public int hashCode() {
-        int hash = 7;
-        hash = 89 * hash + Objects.hashCode(this.title);
-        hash = 89 * hash + Objects.hashCode(this.startDate);
-        hash = 89 * hash + Objects.hashCode(this.completedDate);
-        hash = 89 * hash + (int) (Double.doubleToLongBits(this.amount) ^ (Double.doubleToLongBits(this.amount) >>> 32));
-        hash = 89 * hash + Objects.hashCode(this.description);
-        hash = 89 * hash + Objects.hashCode(this.type);
-        return hash;
-    }
+	@Override
+	public int hashCode() {
+		int hash = 7;
+		hash = 89 * hash + Objects.hashCode(this.title);
+		hash = 89 * hash + Objects.hashCode(this.startDate);
+		hash = 89 * hash + Objects.hashCode(this.completedDate);
+		hash = 89 * hash + (int) (Double.doubleToLongBits(this.amount) ^ (Double.doubleToLongBits(this.amount) >>> 32));
+		hash = 89 * hash + Objects.hashCode(this.description);
+		hash = 89 * hash + Objects.hashCode(this.type);
+		return hash;
+	}
 }
